Face_Swap/requirements.txt
albumentations==1.3.0
appnope==0.1.3
backcall==0.2.0
boto3==1.26.3
botocore==1.29.3
charset-normalizer==2.1.1
coloredlogs==15.0.1
cycler==0.11.0
Cython==0.29.32
debugpy==1.6.3
decorator==5.1.1
easydict==1.10
entrypoints==0.4
facer @ git+https://github.com/FacePerceiver/facer.git@cbb01dc457f3713050e89af7b2c9c0d98663842c
flatbuffers==22.10.26
fonttools==4.38.0
humanfriendly==10.0
idna==3.4
imageio==2.22.3
importlib-metadata==5.0.0
imutils==0.5.4
insightface==0.6.2
ipykernel==6.16.2
ipython==7.34.0
ipywidgets==8.0.2
jedi==0.18.1
jmespath==1.0.1
joblib==1.2.0
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterlab-widgets==3.0.3
kafka-python==2.0.2
kiwisolver==1.4.4
matplotlib==3.5.3
matplotlib-inline==0.1.6
mpmath==1.2.1
nest-asyncio==1.5.6
networkx==2.6.3
numpy==1.21.6
onnx==1.12.0
onnxruntime==1.13.1
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
prettytable==3.5.0
prompt-toolkit==3.0.32
protobuf==3.20.1
psutil==5.9.3
ptyprocess==0.7.0
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==24.0.1
qudida==0.0.4
requests==2.28.1
s3transfer==0.6.0
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sympy==1.10.1
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==1.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.12
validators==0.20.0
wcwidth==0.2.5
widgetsnbextension==4.0.3
zipp==3.10.0
